feat(print-report): give the first page to the map
The masthead keeps the location name alone and the attribution and generation
date close the report on its last page, where they read as a colophon rather
than as a subtitle. `@page :first` narrows the top margin to 10mm, since the
opening page has no running content above it to clear, and the map takes the
space all of that frees: 75mm to 105mm.

The corner mark is the favicon artwork in brand-800 on transparent ground —
`logo-bg.png` carries a teal wedge that prints as a block of ink on every page —
smaller and lifted so it sits level with the title instead of over the map. It
is served unoptimized: re-encoding a 2KB mark buys nothing, and the optimizer
cache kept handing back the previous artwork.
