+++ src/containers/print-report/print-header.tsx
@@ -4,20 +4,12 @@ import { useCallback, useEffect, useMemo, useRef } from 'react';
 
 import { useSearchParams } from 'next/navigation';
 
-import useIsClient from 'hooks/use-is-client';
 import { useSyncLocation } from 'hooks/use-sync-location';
 
 import { useLocation } from '@/containers/datasets/locations/hooks';
 import type { LocationTypes } from '@/containers/datasets/locations/types';
 
 import usePrintReady from './use-print-ready';
-
-const LOCATION_TYPE_LABELS: Partial<Record<LocationTypes | 'worldwide', string>> = {
-  country: 'Country',
-  wdpa: 'Protected area',
-  'custom-area': 'Custom area',
-  worldwide: 'Worldwide',
-};
 
 const PrintHeader = () => {
   const { type, id } = useSyncLocation();
@@ -38,15 +30,6 @@ const PrintHeader = () => {
     return locationName || 'Worldwide';
   }, [locationType, locationName]);
 
-  const typeLabel = LOCATION_TYPE_LABELS[locationType];
-
-  // Client-only: the server has no notion of the reader's locale or timezone,
-  // so formatting the date during SSR would trip a hydration mismatch.
-  const isClient = useIsClient();
-  const generatedOn = isClient
-    ? new Intl.DateTimeFormat(undefined, { dateStyle: 'long' }).format(new Date())
-    : null;
-
   const handlePrint = useCallback(() => {
     window.print();
   }, []);
@@ -62,7 +45,7 @@ const PrintHeader = () => {
     // print button sits outside the flow so it cannot shift that centring.
     // Tighter in print: the first page has to hold the map and the lead widget
     // under it, and every millimetre the masthead takes comes out of the map.
-    <div className="relative py-6 text-center print:py-3">
+    <div className="relative py-6 text-center print:py-2">
       <button
         type="button"
         onClick={handlePrint}
@@ -72,17 +55,11 @@ const PrintHeader = () => {
       >
         {isReady ? 'Print report' : 'Preparing report…'}
       </button>
+      {/* The location alone opens the report — attribution and the generation
+          date close it, on the last page (see ./print-footer). */}
       <h1 className="text-5xl leading-tight font-light text-black/85 first-letter:uppercase print:text-4xl">
         {displayName}
       </h1>
-      <p className="mt-2 text-lg font-light text-black/85">
-        Powered by Global Mangrove Watch &middot;{' '}
-        <span className="notranslate">globalmangrovewatch.org</span>
-      </p>
-      <p className="mt-1 text-sm text-black/54">
-        {typeLabel && <span>{typeLabel} &middot; </span>}
-        {generatedOn && <span className="notranslate">{generatedOn}</span>}
-      </p>
     </div>
   );
 };
